trying to fix lowercase naming problem
committing Ratiw\Commandbus and using it as Ratiw\CommandBus on Windows
does not have any problem. But when install it on Ubuntu on VPS this
causes problem as Ratiw\CommandBus is not the same as Ratiw\Commandbus.
This results in 'Class xxxx not found' by composer's autoload. So, I
have to rename it to something else (like CB) and renaming it back to
CommandBus (the one that I used inside my codes).